requirements.txt
pandas==2.2.3
polars==1.27.0
pyarrow==19.0.1
requests==2.32.3
feedparser==6.0.11
pandera==0.22.1
pydantic==2.11.1
pyyaml==6.0.2
tenacity==9.1.2
duckdb==1.2.2
tqdm==4.67.1
loguru==0.7.3
pytest==8.3.5
python-dotenv==1.1.0